requirements.txt
streamlit
langchain-openai
langchain
python-dotenv
redis